Factor out complex logic for window formatting into shell script
I could not for the life of me get the conditionally formatting working
right using tmux's formatting primitives.  Instead I factored it out
into a shell script and now it works like a charm.

If the key binding table is set to the one used to redirect keys to the
remote tmux instance, the list of windows does not appear at all.  This
helps me to avoid getting confused and trying to run local window
commands